I must fix the AI code after I've verified that it works
Also addressed issue #30

+++ src/ai/chum_tests.rs
@@ -13,10 +13,10 @@ fn test_get_rock_positions_1() {
     let mut board = Board::new_board(Tile::Black, Tile::White);
     board.set_state(&state);
 
-    let mut target_w: Vec<(usize, usize)> = Vec::new();
+    let mut target_w: Vec<(i8, i8)> = Vec::new();
     target_w.push((3, 0));
 
-    let mut target_b: Vec<(usize, usize)> = Vec::new();
+    let mut target_b: Vec<(i8, i8)> = Vec::new();
     target_b.push((0, 3));
 
     assert_eq!(ChumBucket::get_rock_positions(&board, Tile::White), target_w);
@@ -34,13 +34,13 @@ fn test_get_rock_positions_2() {
     let mut board = Board::new_board(Tile::Black, Tile::White);
     board.set_state(&state);
 
-    let mut target_w: Vec<(usize, usize)> = Vec::new();
+    let mut target_w: Vec<(i8, i8)> = Vec::new();
     target_w.push((0, 0));
     target_w.push((0, 1));
     target_w.push((0, 2));
     target_w.push((0, 3));
 
-    let mut target_b: Vec<(usize, usize)> = Vec::new();
+    let mut target_b: Vec<(i8, i8)> = Vec::new();
     target_b.push((3, 0));
     target_b.push((3, 1));
     target_b.push((3, 2));
